load lemmatization lookups during sqlite sync
